refactored to clean it up
