+++ client/src/index.js
@@ -1,63 +1,21 @@
 import io from 'socket.io-client';
 import * as Rx from 'rx';
-import { Game } from 'phaser';
+import { Phaser } from 'phaser';
 
-console.log(PIXI, Phaser, Game);
-console.log(Rx);
+import connectToServer from './js/connectToServer';
+import monitorConnection from './js/monitorConnection';
+import createGame from './js/createGame';
+import gameLogic from './js/gameLogic';
 
-function connectToServer() {
-    const janken_endpoint = 'http://localhost:9000';
-    const socket = io.connect(janken_endpoint);
-    const source = new Rx.Subject();
-
-    socket.on('connect', () => {
-        console.log("connected");
-    });
-
-    socket.on('connect_failed', () => {
-        console.log("connect_failed");
-        socket.close();
-    });
-
-    socket.on('event', (data) => source.onNext(data));
-
-    socket.on('disconnect', () => {
-        console.log("disconnected");
-        socket.close();
-        source.return();
-    });
-
-    socket.emit('command', {hello: 'world'})
-
-    return source;
-}
-
-function createGame() {
-    const game = new Phaser.Game(750, 1334, Phaser.AUTO, '', { preload: preload, create: create, update: update });
-
-    function preload() {
-    }
-
-    function create() {
-        game.scale.scaleMode = Phaser.ScaleManager.NO_SCALE;
-        game.scale.pageAlignVertically = true;
-        game.scale.pageAlignHorizontally = true;
-    }
-
-    function update() {
-    }    
-}
+console.log(PIXI, Phaser, Game, Rx);
 
 window.onload = () => {
     // CONNECT TO SERVER
     const source = connectToServer();
-    const observer = Rx.Observer.create(
-        (x) => console.log('Next', x),
-        (err) => console.log('Error', err),
-        () => console.log('Completed')
-    );
-    const subscription = source.subscribe(observer);
-
+    // MONITOR CONNECTION
+    monitorConnection(source);
+    // START GAME LOGIC
+    gameLogic(source);
     // CREATE GAME CANVAS
     createGame();
 };
